+++ src/argue.rs
@@ -6,6 +6,8 @@ use crate::{
 	ArgyleError,
 	ArgsOsStr,
 	KeyKind,
+	Options,
+	OptionsOsStr,
 };
 use std::{
 	cell::Cell,
@@ -478,9 +480,7 @@ impl Argue {
 	/// let mut args = Argue::new(0).unwrap();
 	/// let switch: bool = args.switch(b"--my-switch");
 	/// ```
-	pub fn switch(&self, key: &[u8]) -> bool {
-		self.args.iter().any(|x| x == key)
-	}
+	pub fn switch(&self, key: &[u8]) -> bool { self.args.iter().any(|x| x == key) }
 
 	#[must_use]
 	#[inline]
@@ -500,6 +500,40 @@ impl Argue {
 	/// ```
 	pub fn switch2(&self, short: &[u8], long: &[u8]) -> bool {
 		self.args.iter().any(|x| x == short || x == long)
+	}
+
+	#[must_use]
+	/// # Switch Count.
+	///
+	/// Return the total number times the switch was specified.
+	///
+	/// ## Examples
+	///
+	/// ```no_run
+	/// use argyle::Argue;
+	///
+	/// let mut args = Argue::new(0).unwrap();
+	/// let count: usize = args.switch_count(b"--my-switch");
+	/// ```
+	pub fn switch_count(&self, key: &[u8]) -> usize {
+		self.args.iter().filter(|&x| x == key).count()
+	}
+
+	#[must_use]
+	/// # Switch x2 Count.
+	///
+	/// Return the total number times the switch was specified.
+	///
+	/// ## Examples
+	///
+	/// ```no_run
+	/// use argyle::Argue;
+	///
+	/// let mut args = Argue::new(0).unwrap();
+	/// let switch2_count: usize = args.switch2_count(b"-s", b"--my-switch");
+	/// ```
+	pub fn switch2_count(&self, short: &[u8], long: &[u8]) -> usize {
+		self.args.iter().filter(|&x| x == short || x == long).count()
 	}
 
 	#[must_use]
@@ -579,13 +613,15 @@ impl Argue {
 	/// # Option.
 	///
 	/// Return the value corresponding to `key`, if present. "Value" in this
-	/// case means the entry immediately following the key. Multi-value
-	/// options are not supported.
+	/// case means the entry immediately following the key.
 	///
 	/// Note: this method is the only way `Argue` knows whether or not a key
 	/// is an option (with a value) or a switch. Be sure to request all
 	/// possible options *before* requesting the trailing arguments to ensure
 	/// the division between named and trailing is properly set.
+	///
+	/// This will only ever match the _last_ occurrence. For options that
+	/// may be specified more than once, use [`Argue::option_values`] instead.
 	///
 	/// ## Examples
 	///
@@ -596,7 +632,7 @@ impl Argue {
 	/// let opt: Option<&[u8]> = args.option(b"--my-opt");
 	/// ```
 	pub fn option(&self, key: &[u8]) -> Option<&[u8]> {
-		let idx = self.args.iter().position(|x| x == key)? + 1;
+		let idx = self.args.iter().rposition(|x| x == key)? + 1;
 		self._option(idx)
 	}
 
@@ -605,6 +641,9 @@ impl Argue {
 	/// This is a convenience method that checks for the existence of two
 	/// options at once, returning the first found value, if any. Generally
 	/// you would use this for a flag that has both a long and short version.
+	///
+	/// This will only ever match the _last_ occurrence. For options that
+	/// may be specified more than once, use [`Argue::option2_iter`] instead.
 	///
 	/// ## Examples
 	///
@@ -615,8 +654,77 @@ impl Argue {
 	/// let opt: Option<&[u8]> = args.option2(b"-o", b"--my-opt");
 	/// ```
 	pub fn option2(&self, short: &[u8], long: &[u8]) -> Option<&[u8]> {
-		let idx = self.args.iter().position(|x| x == short || x == long)? + 1;
+		let idx = self.args.iter().rposition(|x| x == short || x == long)? + 1;
 		self._option(idx)
+	}
+
+	/// # Option Value(s) Iterator.
+	///
+	/// Return any and all values corresponding to `key` (meaning the entries
+	/// immediately following each instance of `key`).
+	///
+	/// This is useful for programs that accept the same flag multiple times,
+	/// or those expecting a byte-delimited value, like a comma-separated list.
+	///
+	/// When using `delimiter`, each value will be carved up by the specified
+	/// byte and returned separately.
+	///
+	/// ## Examples
+	///
+	/// ```no_run
+	/// use argyle::Argue;
+	///
+	/// let mut args = Argue::new(0).unwrap();
+	/// for v in args.option_values(b"--my-option", None) {
+	///     println!("{:?}", std::str::from_utf8(v));
+	/// }
+	/// ```
+	pub fn option_values<'a>(&'a self, key: &'a [u8], delimiter: Option<u8>)
+	-> Options<'a> {
+		if let Some(idx) = self.args.iter().rposition(|x| x == key) {
+			let idx = idx + 1;
+			if idx < self.args.len() {
+				if self.last.get() < idx { self.last.set(idx); }
+				return Options::new(&self.args[..=idx], key, None, delimiter);
+			}
+		}
+
+		Options::default()
+	}
+
+	/// # Option 2x Value(s) Iterator.
+	///
+	/// Return any and all values corresponding to either the `short` or `long`
+	/// version of a key (meaning the entries immediately following each
+	/// instance of them).
+	///
+	/// This is useful for programs that accept the same flag multiple times,
+	/// or those expecting a byte-delimited value, like a comma-separated list.
+	///
+	/// When using `delimiter`, each value will be carved up by the specified
+	/// byte and returned separately.
+	///
+	/// ## Examples
+	///
+	/// ```no_run
+	/// use argyle::Argue;
+	///
+	/// let mut args = Argue::new(0).unwrap();
+	/// for v in args.option2_iter(b"-o", b"--my-opt", None) {
+	///     println!("{:?}", std::str::from_utf8(v));
+	/// }
+	/// ```
+	pub fn option2_iter<'a>(&'a self, short: &'a [u8], long: &'a [u8], delimiter: Option<u8>)
+	-> Options<'a> {
+		if let Some(idx) = self.args.iter().rposition(|x| x == short || x == long) {
+			let idx = idx + 1;
+			if idx < self.args.len() {
+				if self.last.get() < idx { self.last.set(idx); }
+				return Options::new(&self.args[..=idx], short, Some(long), delimiter);
+			}
+		}
+
+		Options::default()
 	}
 
 	#[must_use]
@@ -770,6 +878,24 @@ impl Argue {
 	/// an [`OsStr`](std::ffi::OsStr) instead of bytes.
 	pub fn option2_os(&self, short: &[u8], long: &[u8]) -> Option<&OsStr> {
 		self.option2(short, long).map(OsStr::from_bytes)
+	}
+
+	/// # Option Value(s) Iterator.
+	///
+	/// This works just like [`Argue::option_values`], except it returns the value
+	/// as an [`OsStr`](std::ffi::OsStr) instead of bytes.
+	pub fn option_values_os<'a>(&'a self, key: &'a [u8], delimiter: Option<u8>)
+	-> OptionsOsStr<'a> {
+		OptionsOsStr(self.option_values(key, delimiter))
+	}
+
+	/// # Option 2x Value(s) Iterator.
+	///
+	/// This works just like [`Argue::option2_iter`], except it returns the value
+	/// as an [`OsStr`](std::ffi::OsStr) instead of bytes.
+	pub fn option2_iter_os<'a>(&'a self, short: &'a [u8], long: &'a [u8], delimiter: Option<u8>)
+	-> OptionsOsStr<'a> {
+		OptionsOsStr(self.option2_iter(short, long, delimiter))
 	}
 
 	#[must_use]